+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## Fixes
 
+- Preserve integer values during YAML request deserialization in [#324](https://github.com/TNO-S3/WuppieFuzz/pull/324)
+
 # v1.5.1 (2026-05-19)
 
 ## Highlights